requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

numpy==1.26.4
pandas==2.2.3
pyyaml==6.0.2

scikit-learn==1.5.2
joblib==1.4.2

mlflow==2.17.2

evidently==0.5.1

prometheus-client==0.21.1
requests==2.32.3
